refactor: :recycle: Moved Restaurant card to separate component file
